update markdown function with sorting
